requirements.txt
playwright>=1.44
pymongo>=4.7
python-dotenv>=1.0
requests>=2.31
certifi>=2024.2.2
groq>=0.9.0
tavily-python>=0.5.0
python-telegram-bot>=21.0